FE/requirements.txt
streamlit
python-docx
docx2pdf
PyPDF2
langchain
langchain_community
pypdf